rising/requirements.txt
GPUtil~=1.4.0
aiohttp~=3.8.3
websockets~=12.0
Pillow~=9.3.0
requests~=2.28.2
qrcode~=7.4.2
python-dotenv~=1.0.0